requirements.txt
# --- Core ---
httpx==0.27.2
SQLAlchemy==2.0.36
PyYAML==6.0.2
python-dotenv==1.0.1
tenacity==9.0.0

# --- Telegram bot + scheduler ---
python-telegram-bot==21.6
APScheduler==3.10.4

# --- Dev / test ---
pytest==8.3.3
pytest-asyncio==0.24.0
respx==0.21.1
